+++ bmgt435_elp/urls.py
@@ -1,4 +1,6 @@
 from django.urls import path
+from django.conf import settings
+from django.conf.urls.static import static
 from .apis import *
 
 
@@ -30,12 +32,19 @@ urlpatterns = [
     path("api/leader-board/paginated", CaseRecordApi.leader_board_paginated ),
 
 
-    path('api/manage/user/import', ManageApi.import_users,),
+    path('api/manage/user/import/semester/<int:semester_id>', ManageApi.import_users,),
     path('api/manage/user/view', ManageApi.view_users,),
+
     path('api/manage/group/create', ManageApi.batch_create_group,),
+    path('api/manage/group/paginated', ManageApi.group_view_paginated,),
+
     path('api/manage/semester/create', ManageApi.create_semester,),
+    path('api/manage/semester/all', ManageApi.get_semesters,),
+
     path('api/manage/system/status', ManageApi.system_status,),
 
     
     path('api/feedback/post', FeedbackApi.post, ),
-]
+
+    
+] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT) # for serving static files in development
