+++ sim_server_django/settings.py
@@ -1,13 +1,6 @@
 """
 Django settings for sim_server_django project.
 
-Generated by 'django-admin startproject' using Django 4.2.1.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.1/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.1/ref/settings/
 """
 
 from pathlib import Path
@@ -17,30 +10,26 @@ import os
 BASE_DIR = Path(__file__).resolve().parent.parent
 
 
-# Quick-start development settings - unsuitable for production
-# See https://docs.djangoproject.com/en/4.1/howto/deployment/checklist/
-
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = 'django-insecure-7+pi0)&orq-c)01-0vt$^=^jhs-m$)t3b7h-sx!i+#)lwbs^u5'
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+DEBUG = os.environ.get("BMGT435_DEBUG", False)
 if DEBUG:
     print('Base dir is \t', BASE_DIR)
 
 
-ALLOWED_HOSTS = ['app', '127.0.0.1', 'localhost', 'chaconn3.me'] # set so because the frontend use this name to access the backend, which is a docker network alias
-CSRF_TRUSTED_ORIGINS = ['http://127.0.0.1', os.environ.get("BMGT435_INDEX") or 'http://localhost'] # set so because the frontend use this name to access the backend, which is a docker network alias
-
-MEDIA_ROOT = BASE_DIR.absolute().as_posix() + '/media/'
-if  DEBUG:
-    print('Media root is \t', MEDIA_ROOT)
+INDEX_HOST = os.environ.get("BMGT435_INDEX")    # host name of the frontend
+ALLOWED_HOSTS = ['app', '127.0.0.1', 'localhost', INDEX_HOST]
+CSRF_TRUSTED_ORIGINS = [
+    'http://127.0.0.1','http://localhost', f'http://{INDEX_HOST}',
+    "https://127.0.0.1", "https://localhost", f"https://{INDEX_HOST}"
+]
 
 
 # Application definition
-
 INSTALLED_APPS = [
-    "bmgt435_platform.apps.BmgtPlatformConfig",
+    "bmgt435_elp.apps.BmgtPlatformConfig",
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -50,8 +39,8 @@ INSTALLED_APPS = [
 ]
 
 MIDDLEWARE = [
-    'bmgt435_platform.middlewares.bgmt435Middlewares.CORSMiddleware',
-    'bmgt435_platform.middlewares.bgmt435Middlewares.AuthenticationMiddleware',
+    'bmgt435_elp.middlewares.bgmt435Middlewares.CORSMiddleware',
+    'bmgt435_elp.middlewares.bgmt435Middlewares.AuthenticationMiddleware',
     'django.middleware.security.SecurityMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.common.CommonMiddleware',
@@ -59,7 +48,7 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    # 'bmgt435_platform.middlewares.bgmt435Middlewares.TestModeMiddleware',
+    # 'bmgt435_elp.middlewares.bgmt435Middlewares.TestModeMiddleware',
 ]
 
 ROOT_URLCONF = 'sim_server_django.urls'
@@ -87,34 +76,34 @@ ASGI_APPLICATION = 'sim_server_django.asgi.application'
 # Database
 # https://docs.djangoproject.com/en/4.1/ref/settings/#databases
 
-# DATABASE_ROUTERS = ['bmgt435_platform.utils.databaseUtils.BMGT435_DB_Router', ]
+# DATABASE_ROUTERS = ['bmgt435_elp.utils.databaseUtils.BMGT435_DB_Router', ]
 
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': BASE_DIR / 'db.sqlite3',
-    },
+# DATABASES = {
+#     'default': {
+#         'ENGINE': 'django.db.backends.sqlite3',
+#         'NAME': BASE_DIR / 'db.sqlite3',
+#     },
     # 'analytics': {
     #     'ENGINE': 'django.db.backends.sqlite3',
     #     'NAME': BASE_DIR / 'analytics.sqlite3',
     # }
-}
-
-
-# DATABASES = {
-#     'default': {
-#         'ENGINE': 'django.db.backends.mysql',
-#         'NAME': os.environ.get("BMGT435_MYSQL_DB"),
-#         "HOST":os.environ.get("BMGT435_MYSQL_HOST"),
-#         "PORT":3306,
-#         "USER":os.environ.get("BMGT435_MYSQL_USER"),
-#         "PASSWORD":os.environ.get("BMGT435_MYSQL_PASSWORD"),
-#         'MYSQL': {
-#             'driver': 'pymysql',
-#             'charset': 'utf8mb4',
-#         },
-#     }
 # }
+
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.mysql',
+        'NAME': os.environ.get("BMGT435_MYSQL_DB"),
+        "HOST":os.environ.get("BMGT435_MYSQL_HOST"),
+        "PORT":3306,
+        "USER":os.environ.get("BMGT435_MYSQL_USER"),
+        "PASSWORD":os.environ.get("BMGT435_MYSQL_PASSWORD"),
+        'MYSQL': {
+            'driver': 'pymysql',
+            'charset': 'utf8mb4',
+        },
+    }
+}
 
 
 # Password validation
