+++ sim_server_django/urls.py
@@ -19,6 +19,6 @@ from django.urls import include, path
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("bmgt435/", include("bmgt435_platform.urls")),
-    path("static/", include("bmgt435_platform.urls")),
+    path("bmgt435/", include("bmgt435_elp.urls")),
+    path("static/", include("bmgt435_elp.urls")),
 ]
